Moved helper function into schema
